+++ src/arch/x86_64/CMakeLists.txt
@@ -5,6 +5,7 @@ set(ARCH_OBJECTS
   apic_asm.asm
   apic_longmode.asm
   arch_start.asm
+  arch_musl.cpp
   exceptions.asm
   interrupts.asm
   fiber_asm.asm
